SAD-59 Refactor ModalOptions
